@@ -1,26 +1,23 @@
 {
     "name": "@graph/svg",
     "version": "0.0.0",
+    "type": "module",
     "private": true,
     "exports": {
         ".": "./src/index.ts"
     },
     "scripts": {
-        "test": "echo 'Add test script here'",
-        "lint": "echo 'Add lint script here'"
-    },
-    "dependencies": {
-        "react": "^18.2.0"
+        "test": "vitest",
+        "lint": "eslint ."
     },
     "devDependencies": {
-        "@repo/eslint-config": "workspace:*",
-        "@repo/typescript-config": "workspace:*",
-        "@types/eslint": "^8.56.5",
-        "@types/node": "^20.11.24",
+        "@restruct/eslint-config": "workspace:*",
+        "@restruct/typescript-config": "workspace:*",
         "@types/react": "^18.2.61",
-        "@types/react-dom": "^18.2.19",
-        "eslint": "^8.57.0",
         "typescript": "^5.3.3",
         "vitest": "^2.1.1"
+    },
+    "peerDependencies": {
+        "react": "^18.3.1"
     }
 }
