@@ -1,6 +1,7 @@
 {
     "name": "@structurizr/react",
     "version": "0.0.0",
+    "type": "module",
     "private": true,
     "exports": {
         ".": "./src/index.ts"
@@ -12,20 +13,15 @@
     "dependencies": {
         "@dagrejs/dagre": "^1.1.4",
         "@graph/svg": "workspace:*",
-        "@structurizr/dsl": "workspace:*",
-        "react": "^18.2.0",
-        "uuid": "^10.0.0"
+        "@structurizr/dsl": "workspace:*"
     },
     "devDependencies": {
-        "@repo/eslint-config": "workspace:*",
-        "@repo/typescript-config": "workspace:*",
-        "@types/eslint": "^8.56.5",
-        "@types/node": "^20.11.24",
+        "@restruct/eslint-config": "workspace:*",
+        "@restruct/typescript-config": "workspace:*",
         "@types/react": "^18.2.61",
-        "@types/react-dom": "^18.2.19",
-        "@types/uuid": "^10.0.0",
-        "eslint": "^8.57.0",
-        "typescript": "^5.3.3",
         "vitest": "^2.1.1"
+    },
+    "peerDependencies": {
+        "react": "^18.3.1"
     }
 }
