@@ -1,6 +1,7 @@
 {
-    "name": "@repo/ui",
+    "name": "@restruct/ui",
     "version": "0.0.0",
+    "type": "module",
     "private": true,
     "exports": {
         "./button": "./src/button.tsx",
@@ -9,20 +10,18 @@
     },
     "scripts": {
         "lint": "eslint . --max-warnings 0",
-        "generate:component": "turbo gen react-component"
+        "check-types": "tsc --noEmit"
     },
     "devDependencies": {
-        "@repo/eslint-config": "workspace:*",
-        "@repo/typescript-config": "workspace:*",
+        "@restruct/eslint-config": "workspace:*",
+        "@restruct/typescript-config": "workspace:*",
         "@turbo/gen": "^1.12.4",
-        "@types/node": "^20.11.24",
-        "@types/eslint": "^8.56.5",
         "@types/react": "^18.2.61",
         "@types/react-dom": "^18.2.19",
-        "eslint": "^8.57.0",
-        "typescript": "^5.3.3"
+        "typescript": "~5.6.2"
     },
     "dependencies": {
-        "react": "^18.2.0"
+        "react": "^19.0.0",
+        "react-dom": "^19.0.0"
     }
 }
