@@ -1,9 +1,9 @@
 /** @type {import("eslint").Linter.Config} */
 module.exports = {
-  root: true,
-  extends: ["@repo/eslint-config/next.js"],
-  parser: "@typescript-eslint/parser",
-  parserOptions: {
-    project: true,
-  },
+    root: true,
+    extends: ["@restruct/eslint-config/next.js"],
+    parser: "@typescript-eslint/parser",
+    parserOptions: {
+        project: true,
+    },
 };
