@@ -3,5 +3,7 @@ import { IDiagram } from "./IDiagram";
 import { IPerson } from "./IPerson";
 import { ISoftwareSystem } from "./ISoftwareSystem";
 
-export interface IComponentDiagram
-    extends IDiagram<IContainer, ISoftwareSystem | IContainer | IPerson> {}
+export type IComponentDiagram = IDiagram<
+    IContainer,
+    ISoftwareSystem | IContainer | IPerson
+>;
