@@ -2,5 +2,7 @@ import { IDiagram } from "./IDiagram";
 import { IPerson } from "./IPerson";
 import { ISoftwareSystem } from "./ISoftwareSystem";
 
-export interface IContainerDiagram
-    extends IDiagram<ISoftwareSystem, ISoftwareSystem | IPerson> {}
+export type IContainerDiagram = IDiagram<
+    ISoftwareSystem,
+    ISoftwareSystem | IPerson
+>;
