@@ -1,4 +1,4 @@
 import { IDiagram } from "./IDiagram";
 import { IModel } from "./IModel";
 
-export interface ISystemLandscapeDiagram extends IDiagram<IModel, unknown> {}
+export type ISystemLandscapeDiagram = IDiagram<IModel, unknown>;
