@@ -31,7 +31,7 @@ export class SystemContextView implements ISupportSnapshot<ISystemContextView> {
     public include: Array<Identifier | All>;
     public exclude: Array<Identifier>;
     public autoLayout?: AutoLayout;
-    public animation?: any;
+    public animation?: unknown;
     public title?: string;
     public description?: string;
     // public properties?: Properties;
@@ -44,7 +44,7 @@ export class SystemContextView implements ISupportSnapshot<ISystemContextView> {
             include: this.include.map((x) => x.toString()),
             exclude: this.exclude.map((x) => x.toString()),
             autoLayout: this.autoLayout?.toSnapshot(),
-            animation: this.animation,
+            animation: {},
             title: this.title,
             description: this.description,
             // properties: this.properties,
