@@ -4,10 +4,14 @@ import {
     ViewType,
     createSystemLandscapeDiagram,
     isPerson,
-    isSoftwareSystem
+    isSoftwareSystem,
 } from "@structurizr/dsl";
 import { FC, PropsWithChildren, useEffect, useState } from "react";
-import { IViewMetadata, ViewMetadataProvider, useWorkspace } from "../../containers";
+import {
+    IViewMetadata,
+    ViewMetadataProvider,
+    useWorkspace,
+} from "../../containers";
 import { ZoomCallback } from "../../types";
 import {
     createDefaultSystemLandscapeView,
@@ -18,56 +22,74 @@ import { Relationship } from "./Relationship";
 import { Person } from "./Person";
 import { Group } from "./Group";
 
-export const SystemLandscapeDiagram: FC<PropsWithChildren<{
-    value?: ISystemLandscapeView;
-    metadata?: IViewMetadata;
-    onZoomInClick?: ZoomCallback;
-    onZoomOutClick?: ZoomCallback;
-}>> = ({
-    children,
-    value,
-    onZoomInClick,
-    onZoomOutClick,
-}) => {
-        const { workspace } = useWorkspace();
-        const [diagram, setDiagram] = useState<ISystemLandscapeDiagram | null>(null);
-        const [metadata, setMetadata] = useState<IViewMetadata>({ elements: {}, relationships: {} });
+export const SystemLandscapeDiagram: FC<
+    PropsWithChildren<{
+        value?: ISystemLandscapeView;
+        metadata?: IViewMetadata;
+        onZoomInClick?: ZoomCallback;
+        onZoomOutClick?: ZoomCallback;
+    }>
+> = ({ children, value, onZoomInClick, onZoomOutClick }) => {
+    const { workspace } = useWorkspace();
+    const [diagram, setDiagram] = useState<ISystemLandscapeDiagram | null>(
+        null
+    );
+    const [metadata, setMetadata] = useState<IViewMetadata>({
+        elements: {},
+        relationships: {},
+    });
 
-        useEffect(() => {
-            if (workspace) {
-                const systemLandscapeView = [workspace.views.systemLandscape].find(x => x?.key === value?.key)
-                    ?? createDefaultSystemLandscapeView();
+    useEffect(() => {
+        if (workspace) {
+            const systemLandscapeView =
+                [workspace.views.systemLandscape].find(
+                    (x) => x?.key === value?.key
+                ) ?? createDefaultSystemLandscapeView();
 
-                const diagram = createSystemLandscapeDiagram(workspace, systemLandscapeView);
-                setDiagram(diagram);
+            const diagram = createSystemLandscapeDiagram(
+                workspace,
+                systemLandscapeView
+            );
+            setDiagram(diagram);
 
-                const metadataAuto = autolayoutDiagram(diagram, ViewType.SystemLandscape);
-                setMetadata(metadataAuto);
-            }
-        }, [workspace, onZoomInClick, onZoomOutClick]);
+            const metadataAuto = autolayoutDiagram(
+                diagram,
+                ViewType.SystemLandscape
+            );
+            setMetadata(metadataAuto);
+        }
+    }, [workspace, onZoomInClick, onZoomOutClick, value?.key]);
 
-        return (
-            <ViewMetadataProvider metadata={metadata} setMetadata={setMetadata}>
-                {diagram?.scope.groups.map((group) => (
-                    <Group key={group.identifier} value={group}>
-                        {group.people.filter(isPerson).map((element) => (
-                            <Person key={element.identifier} value={element} />
+    return (
+        <ViewMetadataProvider metadata={metadata} setMetadata={setMetadata}>
+            {diagram?.scope.groups.map((group) => (
+                <Group key={group.identifier} value={group}>
+                    {group.people.filter(isPerson).map((element) => (
+                        <Person key={element.identifier} value={element} />
+                    ))}
+                    {group.softwareSystems
+                        .filter(isSoftwareSystem)
+                        .map((element) => (
+                            <SoftwareSystem
+                                key={element.identifier}
+                                value={element}
+                            />
                         ))}
-                        {group.softwareSystems.filter(isSoftwareSystem).map((element) => (
-                            <SoftwareSystem key={element.identifier} value={element} />
-                        ))}
-                    </Group>
-                ))}
-                {diagram?.scope.people.map((element) => (
-                    <Person key={element.identifier} value={element} />
-                ))}
-                {diagram?.scope.softwareSystems.map((element) => (
-                    <SoftwareSystem key={element.identifier} value={element} />
-                ))}
-                {diagram?.relationships.map((relationship) => (
-                    <Relationship key={relationship.identifier} value={relationship} />
-                ))}
-                {children}
-            </ViewMetadataProvider>
-        );
-    };
+                </Group>
+            ))}
+            {diagram?.scope.people.map((element) => (
+                <Person key={element.identifier} value={element} />
+            ))}
+            {diagram?.scope.softwareSystems.map((element) => (
+                <SoftwareSystem key={element.identifier} value={element} />
+            ))}
+            {diagram?.relationships.map((relationship) => (
+                <Relationship
+                    key={relationship.identifier}
+                    value={relationship}
+                />
+            ))}
+            {children}
+        </ViewMetadataProvider>
+    );
+};
